backend/requirements.txt
asgiref==3.6.0
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Markdown==3.3.4
psycopg2-binary==2.9.5
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
uWSGI==2.0.21
webencodings==0.5.1
django-cors-headers==4.1.0
python-dotenv==0.19.2
djangorestframework-simplejwt
whitenoise==6.4.0
